pyproject.toml
[project]
name = "nosql-final-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.13,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pymongo (>=4.13.2,<5.0.0)",
    "motor (>=3.7.1,<4.0.0)",
    "neo4j (>=5.28.1,<6.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "streamlit (>=1.46.1,<2.0.0)",
    "certifi (>=2025.6.15,<2026.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
httpx = "^0.28.1"

